Part 2 of fixing Dispose methods of solution

+++ Server/WorldServer/Objects/WS.Items.vb
@@ -595,9 +595,27 @@ Public Module WS_Items
             'DONE: Internal database fixers
             If Stackable = 0 Then Stackable = 1
         End Sub
-        Public Sub Dispose() Implements System.IDisposable.Dispose
-            ITEMDatabase.Remove(Id)
+
+#Region "IDisposable Support"
+        Private disposedValue As Boolean ' To detect redundant calls
+
+        ' IDisposable
+        Protected Overridable Sub Dispose(disposing As Boolean)
+            If Not Me.disposedValue Then
+                ' TODO: free unmanaged resources (unmanaged objects) and override Finalize() below.
+                ' TODO: set large fields to null.
+                ITEMDatabase.Remove(Id)
+            End If
+            Me.disposedValue = True
         End Sub
+
+        ' This code added by Visual Basic to correctly implement the disposable pattern.
+        Public Sub Dispose() Implements IDisposable.Dispose
+            ' Do not change this code.  Put cleanup code in Dispose(ByVal disposing As Boolean) above.
+            Dispose(True)
+            GC.SuppressFinalize(Me)
+        End Sub
+#End Region
 
         'Item's visuals
         Public Model As Integer = 0
@@ -1220,15 +1238,33 @@ Public Module WS_Items
             End If
             Me.Dispose()
         End Sub
-        Public Sub Dispose() Implements System.IDisposable.Dispose
-            WORLD_ITEMs.Remove(GUID)
 
-            If ITEMDatabase(ItemEntry).IsContainer() Then
-                For Each Item As KeyValuePair(Of Byte, ItemObject) In Items
-                    Item.Value.Dispose()
-                Next
+#Region "IDisposable Support"
+        Private disposedValue As Boolean ' To detect redundant calls
+
+        ' IDisposable
+        Protected Overridable Sub Dispose(disposing As Boolean)
+            If Not Me.disposedValue Then
+                ' TODO: free unmanaged resources (unmanaged objects) and override Finalize() below.
+                ' TODO: set large fields to null.
+                WORLD_ITEMs.Remove(GUID)
+
+                If ITEMDatabase(ItemEntry).IsContainer() Then
+                    For Each Item As KeyValuePair(Of Byte, ItemObject) In Items
+                        Item.Value.Dispose()
+                    Next
+                End If
             End If
+            Me.disposedValue = True
         End Sub
+
+        ' This code added by Visual Basic to correctly implement the disposable pattern.
+        Public Sub Dispose() Implements IDisposable.Dispose
+            ' Do not change this code.  Put cleanup code in Dispose(ByVal disposing As Boolean) above.
+            Dispose(True)
+            GC.SuppressFinalize(Me)
+        End Sub
+#End Region
 
         Public Function IsBroken() As Boolean
             Return (Durability = 0) AndAlso (ItemInfo.Durability > 0)
